ppv-platform/fastapi_app/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
python-dotenv
httpx
PyJWT
